packages/converter/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "webp-converter"
version = "0.1.0"
description = "Core WebP conversion logic with cwebp wrapper and image analysis"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0",
    "numpy>=1.24",
    "opencv-python>=4.8",
    "pillow-heif>=0.15",
    "webp-shared",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["webp_converter*"]
